# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AutoGenBook API
  description: HTTP API for AutoGenBook, served by this FastAPI app and reached through
    the nginx-fronted Docker Compose stack. Every route is mounted under /api/v1;
    there is no authentication and no CORS middleware. Error responses use RFC 9457
    problem details (application/problem+json); list endpoints return a {items, total,
    limit, offset} page envelope.
  version: 0.1.0
paths:
...
  /api/v1/projects/{project_id}/runs:
    post:
      tags:
      - runs
      summary: Create Run
      operationId: runs-create_run
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunOptionsIn'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - runs
      summary: List Runs
      operationId: runs-list_runs
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_Run_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/outline/{node_id}/regenerate:
    post:
      tags:
      - runs
      summary: Regenerate Node
      operationId: runs-regenerate_node
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Node Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegenerateRequestIn'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}:
    get:
      tags:
      - runs
      summary: Get Run
      operationId: runs-get_run
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/cancel:
    post:
      tags:
      - runs
      summary: Cancel Run
      operationId: runs-cancel_run
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/exports:
    post:
      tags:
      - runs
      summary: Export Run
      operationId: runs-export_run
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequestIn'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/events:
    get:
      tags:
      - runs
      summary: List Run Events
      operationId: runs-list_run_events
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      - name: afterSeq
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Afterseq
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunEventPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/artifacts:
    get:
      tags:
      - runs
      summary: List Run Artifacts
      operationId: runs-list_run_artifacts
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_RunArtifact_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/events/stream:
    get:
      tags:
      - runs
      summary: Stream Run Events
      operationId: runs-stream_run_events
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    ArtifactKind:
      type: string
      enum:
      - markdown
      - tex
      - pdf
      - structure_graph
      - book_structure
      - section
      - section_review
      - kb_sources
      - run_meta
      - llm_usage
      - audit_report
      - log
      - bib
      - other
      title: ArtifactKind
...
    ExportRequestIn:
      properties:
        format:
          type: string
          enum:
          - latex
          - pdf
          title: Format
      additionalProperties: false
      type: object
      required:
      - format
      title: ExportRequestIn
      description: Request body of `POST /runs/{id}/exports`.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Page_RunArtifact_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RunArtifact'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[RunArtifact]
    Page_Run_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Run'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[Run]
...
    RegenerateRequestIn:
      properties:
        promptModifier:
          anyOf:
          - type: string
          - type: 'null'
          title: Promptmodifier
      additionalProperties: false
      type: object
      title: RegenerateRequestIn
      description: Request body of `POST /projects/{id}/outline/{nodeId}/regenerate`.
    Run:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        projectId:
          type: string
          format: uuid
          title: Projectid
        kind:
          $ref: '#/components/schemas/RunKind'
        status:
          $ref: '#/components/schemas/RunStatus'
        options:
          $ref: '#/components/schemas/RunOptionsOut'
        baseRunId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Baserunid
        targetNodeId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Targetnodeid
        exitCode:
          anyOf:
          - type: integer
          - type: 'null'
          title: Exitcode
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        totalTokens:
          anyOf:
          - type: integer
          - type: 'null'
          title: Totaltokens
        totalCostUsd:
          anyOf:
          - type: number
          - type: 'null'
          title: Totalcostusd
        queuedAt:
          type: string
          format: date-time
          title: Queuedat
        startedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Startedat
        finishedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finishedat
        resumable:
          type: boolean
          title: Resumable
      type: object
      required:
      - id
      - projectId
      - kind
      - status
      - options
      - baseRunId
      - targetNodeId
      - exitCode
      - error
      - totalTokens
      - totalCostUsd
      - queuedAt
      - startedAt
      - finishedAt
      - resumable
      title: Run
    RunArtifact:
      properties:
        kind:
          $ref: '#/components/schemas/ArtifactKind'
        relativePath:
          type: string
          title: Relativepath
        fileId:
          type: string
          format: uuid
          title: Fileid
        filename:
          type: string
          title: Filename
        sizeBytes:
          type: integer
          title: Sizebytes
        contentType:
          type: string
          title: Contenttype
      type: object
      required:
      - kind
      - relativePath
      - fileId
      - filename
      - sizeBytes
      - contentType
      title: RunArtifact
    RunEvent:
      properties:
        seq:
          type: integer
          title: Seq
        ts:
          type: string
          format: date-time
          title: Ts
        level:
          type: string
          title: Level
        stage:
          type: string
          title: Stage
        message:
          type: string
          title: Message
        payload:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Payload
      type: object
      required:
      - seq
      - ts
      - level
      - stage
      - message
      title: RunEvent
    RunEventPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RunEvent'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        afterSeq:
          type: integer
          title: Afterseq
      type: object
      required:
      - items
      - total
      - limit
      - afterSeq
      title: RunEventPage
      description: '`GET /runs/{id}/events`''s response envelope. Deliberately not
        `Page`:

        this endpoint pages by `seq` (a monotonically increasing cursor), not by

        row position, so `Page.offset` echoing `afterSeq` misrepresented a

        sequence cursor as a row offset (issue #61).'
    RunKind:
      type: string
      enum:
      - full
      - regenerate_section
      - export
      title: RunKind
    RunOptionsIn:
      properties:
        outline:
          type: string
          enum:
          - project
          - generate
          title: Outline
          default: project
        outputFormat:
          anyOf:
          - type: string
            enum:
            - markdown
            - latex
            - pdf
          - type: 'null'
          title: Outputformat
        allowSubdivision:
          type: boolean
          title: Allowsubdivision
          default: false
        enableWebRag:
          type: boolean
          title: Enablewebrag
          default: false
        auditBook:
          type: boolean
          title: Auditbook
          default: false
        auditBookMode:
          type: string
          enum:
          - 'off'
          - warn
          - strict
          title: Auditbookmode
          default: warn
        legacyTex:
          type: boolean
          title: Legacytex
          default: false
        rebuildKb:
          type: boolean
          title: Rebuildkb
          default: false
        failFastSchema:
          type: boolean
          title: Failfastschema
          default: false
        resume:
          type: boolean
          title: Resume
          default: false
        exportTexOnly:
          type: boolean
          title: Exporttexonly
          default: false
      additionalProperties: false
      type: object
      title: RunOptionsIn
      description: 'Request body of `POST /projects/{id}/runs`. `output_format` defaults

        to the project''s own `output_format` when omitted (resolved by

        `RunService.create`''s caller, not here).'
    RunOptionsOut:
      properties:
        outline:
          type: string
          enum:
          - project
          - generate
          title: Outline
        outputFormat:
          type: string
          enum:
          - markdown
          - latex
          - pdf
          title: Outputformat
        allowSubdivision:
          type: boolean
          title: Allowsubdivision
        enableWebRag:
          type: boolean
          title: Enablewebrag
        auditBook:
          type: boolean
          title: Auditbook
        auditBookMode:
          type: string
          enum:
          - 'off'
          - warn
          - strict
          title: Auditbookmode
        legacyTex:
          type: boolean
          title: Legacytex
        rebuildKb:
          type: boolean
          title: Rebuildkb
        failFastSchema:
          type: boolean
          title: Failfastschema
        resume:
          type: boolean
          title: Resume
        exportTexOnly:
          type: boolean
          title: Exporttexonly
        promptModifier:
          anyOf:
          - type: string
          - type: 'null'
          title: Promptmodifier
      type: object
      required:
      - outline
      - outputFormat
      - allowSubdivision
      - enableWebRag
      - auditBook
      - auditBookMode
      - legacyTex
      - rebuildKb
      - failFastSchema
      - resume
      - exportTexOnly
      title: RunOptionsOut
    RunStatus:
      type: string
      enum:
      - queued
      - running
      - succeeded
      - failed
      - cancelled
      title: RunStatus
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError